projects/streamlit/pyproject.toml
[tool.poetry]
name = "deteccao-alagamento"
version = "0.1.0"
description = ""
authors = ["Gabriel Gazola Milan <gabriel.gazola@poli.ufrj.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.28.1"
folium = "0.14.0"
streamlit-folium = "^0.15.1"
streamlit-extras = "^0.3.5"
streamlit-autorefresh = "^1.0.1"
pillow = "^10.1.0"
streamlit-aggrid = "0.3.4.post3"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.flake8]
max-line-length = 100
exclude = [".git", "__pycache__", "dist"]
max-complexity = 10

[tool.isort]
atomic = true
profile = "black"
line_length = 100
skip_gitignore = true